backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
onnx==1.17.0
onnxruntime==1.19.2
numpy==2.1.1
pydantic==2.9.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1